Rename libcloud.resources.lb to libcloud.loadbalancer'
git-svn-id: https://svn.apache.org/repos/asf/incubator/libcloud/trunk@1102975 13f79535-47bb-0310-9956-ffa450edef68
