+++ test/loadbalancer/fixtures/rackspace/v1_slug_loadbalancers_8290.json
