+++ test/loadbalancer/fixtures/rackspace/v1_slug_loadbalancers_8290_nodes_post.json
