+++ test/loadbalancer/fixtures/rackspace/v1_slug_loadbalancers_post.json
