+++ test/loadbalancer/test_rackspace.py
@@ -3,11 +3,11 @@ import os.path
 import sys
 import unittest
 
-from libcloud.resource.lb.base import LB, LBNode
-from libcloud.resource.lb.drivers.rackspace import RackspaceLBDriver
+from libcloud.loadbalancer.base import LB, LBNode
+from libcloud.loadbalancer.drivers.rackspace import RackspaceLBDriver
 
 from test import MockHttp, MockRawResponse
-from test.file_fixtures import ResourceFileFixtures
+from test.file_fixtures import LoadBalancerFileFixtures
 
 class RackspaceLBTests(unittest.TestCase):
 
@@ -72,7 +72,7 @@ class RackspaceLBTests(unittest.TestCase):
         self.assertTrue(ret)
 
 class RackspaceLBMockHttp(MockHttp):
-    fixtures = ResourceFileFixtures(os.path.join('lb', 'rackspace'))
+    fixtures = LoadBalancerFileFixtures('rackspace')
 
     def _v1_0(self, method, url, body, headers):
         headers = {'x-server-management-url': 'https://servers.api.rackspacecloud.com/v1.0/slug',
